requirements.txt
streamlit
opencv-python-headless 
numpy
pytesseract
tensorflow
Pillow
matplotlib
scikit-learn
